Move spatial utilities into separate module
